requirements.txt
fastapi
uvicorn
pycryptodome
openai
groq